+++ Source/Modules/app_models/Sequences/StepChannel.h
@@ -18,9 +18,16 @@ class StepChannel {
     juce::BigInteger getPattern();
     void setNote(int noteIndex, bool value);
     bool getNote(int noteIndex);
+    /**
+     * Maximum of notes based on notes per measure
+     * Initially we wanted 4 notes in 4 measures but with if we increase the
+     * number of notes per measure we get only 1 measure. To have 4 measures
+     * with 16 notes per measure we need 64.
+     */
+    static int getMaxNumberOfNotes(int notesPerMeasure);
 
     static const int maxNumberOfChannels;
-    static const int maxNumberOfNotes;
+    static const int maxNumberOfMeasures;
 
     // State must be public for value tree object list
     juce::ValueTree state;